+++ tests/fields/test_complex_base_field.py
@@ -1,0 +1,10 @@
+import pytest
+
+from mongoengine.base import ComplexBaseField
+from tests.utils import MongoDBTestCase
+
+
+class TestComplexBaseField(MongoDBTestCase):
+    def test_field_validation(self):
+        with pytest.raises(TypeError, match="field argument must be a Field instance"):
+            ComplexBaseField("test")